Move thumbnail code into a ValueConverter
